requirements.txt
agno==1.7.5
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.1
attrs==25.3.0
Authlib==1.6.1
babel==2.17.0
backoff==2.2.1
baidusearch==1.0.3
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
brave-search==0.1.8
Brotli==1.1.0
browser-use==0.5.5
bs4==0.0.2
bubus==1.4.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
courlan==1.3.2
Crawl4AI==0.6.3
cryptography==45.0.4
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==2025.4.1
ecdsa==0.19.1
email_validator==2.2.0
exa-py==1.10.0
exceptiongroup==1.2.0
executing==2.0.1
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
firecrawl==2.16.1
firecrawl-py==2.16.1
Flask==2.3.3
Flask-Login==0.6.3
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-genai==1.25.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.1.1
groq==0.30.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
holidays==0.80
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanize==4.12.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.4
ipython==8.22.2
ipython_pygments_lexers==0.3.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jusText==3.0.2
kiwisolver==1.4.9
lancedb==0.21.2
langchain==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.21
litellm==1.73.6
lxml==5.3.1
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
markdown_pdf==1.7
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.12.1
mdurl==0.1.2
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
narwhals==1.33.0
nest_asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
ollama==0.5.1
openai==1.102.0
orjson==3.11.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
passlib==1.7.4
patchright==1.52.5
peewee==3.17.9
pgvector==0.4.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.7
playwright==1.53.0
plotly==6.2.0
pluggy==1.5.0
portalocker==2.10.1
posthog==6.3.1
primp==0.14.0
prompt_toolkit==3.0.43
propcache==0.3.2
prophet==1.1.7
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg==3.2.6
pure_eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.6.0
PyMuPDF==1.25.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.8.0
pyperclip==1.9.0
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.3.5
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==25.1.2
rank-bm25==0.2.2
readability-lxml==0.8.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
scrapegraph_py==1.14.2
screeninfo==0.8.1
seaborn==0.13.2
sentence-transformers==5.1.0
sentry-sdk==2.35.0
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.4.1
stack_data==0.6.3
stanio==0.5.1
starlette==0.47.3
streamlit==1.44.1
sympy==1.14.0
tantivy==0.22.2
tenacity==8.5.0
textblob==0.19.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.1.3
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.55.0
twilio==9.7.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==1.26.20
uuid7==0.1.0
uvicorn==0.35.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yahoo-fin==0.8.9.1
yarl==1.20.1
yfinance==0.2.65
zipp==3.21.0
zstandard==0.24.0